pyproject.toml
[project]
name = "ai-agent-setup"
version = "0.1.0"
description = "AI Agent setup for automating digital tasks"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "openai>=1.68.2",
    "pydantic>=2.10.6",
    "python-dotenv>=1.0.1",
]
